+++ app/pb_kits/playbook/pb_legend/docs/_legend_prefix.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Legend } from '../..'
+
+const LegendPrefix = () => (
+  <div>
+    <Legend
+        color="data_3"
+        prefixText="10"
+        text="Windows"
+    />
+  </div>
+)
+
+export default LegendPrefix